+++ packages/eslint-plugin/utils/index.js
@@ -1,8 +1,13 @@
-const { TRANSLATION_FUNCTIONS, REGEXP_PLACEHOLDER } = require( './constants' );
+const {
+	TRANSLATION_FUNCTIONS,
+	REGEXP_SPRINTF_PLACEHOLDER,
+	REGEXP_SPRINTF_PLACEHOLDER_UNORDERED,
+} = require( './constants' );
 const { getTranslateStrings } = require( './get-translate-strings' );
 
 module.exports = {
 	TRANSLATION_FUNCTIONS,
-	REGEXP_PLACEHOLDER,
+	REGEXP_SPRINTF_PLACEHOLDER,
+	REGEXP_SPRINTF_PLACEHOLDER_UNORDERED,
 	getTranslateStrings,
 };